+++ tests/AutoBot.Tests/Services/ProcessTradeExecutionTests.cs
@@ -575,11 +575,11 @@ public class ProcessTradeExecutionTests
         decimal exitPrice,
         decimal leverage,
         TradeSide side,
-        decimal expectedMargin,
+        long expectedMargin,
         decimal expectedLiquidation,
-        decimal expectedOpeningFee,
-        decimal expectedClosingFee,
-        decimal expectedPL)
+        long expectedOpeningFee,
+        long expectedClosingFee,
+        long expectedPL)
     {
         // Act
         var trade = TradeFactory.CreateTrade(
@@ -591,13 +591,13 @@ public class ProcessTradeExecutionTests
             TradeState.Open);
 
         // Assert
-        Assert.Equal(quantity, trade.quantity);
-        Assert.Equal(entryPrice, trade.price);
+        Assert.Equal(quantity, (decimal)trade.quantity);
+        Assert.Equal(entryPrice, (decimal)trade.price);
         Assert.Equal(leverage, trade.leverage);
         Assert.Equal(side.ToString().ToLower(), trade.side);
         Assert.Equal(expectedMargin, trade.margin);
         Assert.Equal(expectedPL, trade.pl);
-        Assert.Equal(expectedLiquidation, trade.liquidation);
+        Assert.Equal(expectedLiquidation, (decimal)trade.liquidation);
         Assert.Equal(expectedOpeningFee, trade.opening_fee);
         Assert.Equal(expectedClosingFee, trade.closing_fee);
     }
